big reorganise / remove plugins/vf-[plugin] requirement
