--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-data-resources.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Data Resources
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/data-resources', __('Data resources'));
