--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-example.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Example
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/example', __('Example'));
