--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-factoid.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Factoid
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/factoid', __('Factoid'));
