--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-group-header.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Group Header
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/group-header', __('Group Header'));
