--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-jobs.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Jobs
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/jobs', __('Jobs'));
