--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-latest-posts.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Latest Posts
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/latest-posts', __('Latest Posts'));
