--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-members.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Members
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/members', __('Members'));
