--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-publications-group-ebi.jsx
@@ -1,11 +1,0 @@
-/**
- * VF-WP EBI Team publications
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings(
-  'vf/publications-group-ebi',
-  __('EBI Team publications')
-);
