--- wp-content/plugins/vf-gutenberg/blocks/plugins/vf-publications.jsx
@@ -1,8 +1,0 @@
-/**
- * VF-WP Publications
- */
-import useSettings from './settings';
-
-const {__} = wp.i18n;
-
-export default useSettings('vf/publications', __('Publications'));
